Completes CRUD for Course

+++ app/Http/Controllers/CourseController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\CourseOffered;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -30,7 +31,7 @@ class CourseController extends Controller
     public function add(Request $request)
     {
         $this->validate($request, Course::$rules);
-        return response()->json(Course::create($request->all(), Response::HTTP_CREATED));
+        return response()->json(Course::create($request->all()), Response::HTTP_CREATED);
     }
 
     public function update(Request $request, $id)
@@ -40,10 +41,10 @@ class CourseController extends Controller
             return response([],Response::HTTP_NOT_FOUND);
         }
         $this->validate($request, Course::$rules);
-        return response()->json(
-            $course->update($request->all()),
-            Response::HTTP_OK
-        );
+        if ($course->update($request->all())) {
+          return response()->json(Course::find($id),Response::HTTP_OK);
+        }
+        return response()->json(['message' => 'Unknown error while updating the course.'],Response::HTTP_CONFLICT);
     }
 
     public function delete($id)
@@ -52,8 +53,12 @@ class CourseController extends Controller
         if (is_null($course)) {
             return response([],Response::HTTP_NOT_FOUND);
         }
+        $course_is_offered = sizeof($course->usages()) > 0;
+        if ($course_is_offered) {
+           return response()->json(['message' => 'Offered courses cannot be deleted.'], Response::HTTP_CONFLICT);
+        }
         Course::destroy($id);
-        return response([],Response::HTTP_OK);
+        return response()->json($course,Response::HTTP_OK);
     }
 
     public function search(Request $request) {
